feat: delete playlist and song from playlist

+++ src/api/playlists/handler.js
@@ -12,6 +12,7 @@ class PlaylistsHandler {
     this.getSongsInPlaylistHandler = this.getSongsInPlaylistHandler.bind(this);
     this.putPlaylistByIdHandler = this.putPlaylistByIdHandler.bind(this);
     this.deletePlaylistByIdHandler = this.deletePlaylistByIdHandler.bind(this);
+    this.deleteSongFromPlaylistHandler = this.deleteSongFromPlaylistHandler.bind(this);
   }
 
   async postPlaylistHandler(request, h) {
@@ -228,11 +229,48 @@ class PlaylistsHandler {
   async deletePlaylistByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyPlaylistAccess(id, credentialId);
       await this._service.deletePlaylistById(id);
 
       return {
         status: 'success',
         message: 'Playlist berhasil dihapus',
+      };
+    } catch (error) {
+      if (error instanceof ClientError) {
+        const response = h.response({
+          status: 'fail',
+          message: error.message,
+        });
+        response.code(error.statusCode);
+        return response;
+      }
+
+      // Server ERROR!
+      const response = h.response({
+        status: 'error',
+        message: 'Maaf, terjadi kegagalan pada server kami.',
+      });
+      response.code(500);
+      console.error(error);
+      return response;
+    }
+  }
+
+  async deleteSongFromPlaylistHandler(request, h) {
+    try {
+      const { id } = request.params;
+      const { songId } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyPlaylistAccess(id, credentialId);
+      await this._service.deleteSongFromPlaylist(id, songId);
+
+      return {
+        status: 'success',
+        message: 'Lagu berhasil dihapus dari Playlist',
       };
     } catch (error) {
       if (error instanceof ClientError) {
